+++ python/GafferOSLUI/OSLCodeUI.py
@@ -106,7 +106,7 @@ Gaffer.Metadata.registerNode(
 
 		"parameters.*" : [
 
-			"labelPlugValueWidget:renameable", True,
+			"renameable", True,
 			# Since the names are used directly as variable names in the code,
 			# it's best to avoid any fancy label formatting for them.
 			"label", lambda plug : plug.getName(),
@@ -133,7 +133,7 @@ Gaffer.Metadata.registerNode(
 
 		"out.*" : [
 
-			"labelPlugValueWidget:renameable", True,
+			"renameable", True,
 			"label", lambda plug : plug.getName(),
 
 		],