+++ python/GafferUI/NodeEditor.py
@@ -67,6 +67,7 @@ class NodeEditor( GafferUI.NodeSetEditor ) :
 		self._doPendingUpdate()
 		return self.__nodeUI
 
+	## \deprecated
 	def setReadOnly( self, readOnly ) :
 
 		if readOnly == self.__readOnly :
@@ -77,6 +78,7 @@ class NodeEditor( GafferUI.NodeSetEditor ) :
 			self.__nodeUI.setReadOnly( readOnly )
 			self.__nameWidget.setEditable( not readOnly )
 
+	## \deprecated
 	def getReadOnly( self ) :
 
 		return self.__readOnly
@@ -124,13 +126,6 @@ class NodeEditor( GafferUI.NodeSetEditor ) :
 
 				GafferUI.Label( "<h4>Node Name</h4>" )
 				self.__nameWidget = GafferUI.NameWidget( node )
-				## \todo Make NameWidget support the readOnly metadata internally itself.
-				# We can't do that easily right now, because it would need to be managing
-				# the exact same `setEditable()` call that we're using here to propagate
-				# our Widget readonlyness. Really our Widget readonlyness mechanism is a
-				# bit lacking, and it should really be inherited automatically so we don't
-				# have to propagate it like this.
-				self.__nameWidget.setEditable( not self.getReadOnly() and not Gaffer.MetadataAlgo.readOnly( node ) )
 
 				with GafferUI.ListContainer(
 					GafferUI.ListContainer.Orientation.Horizontal,