+++ python/GafferUI/NodeUI.py
@@ -59,6 +59,7 @@ Gaffer.Metadata.registerNode(
 	""",
 
 	"documentation:url", __documentationURL,
+	"renameable", True,
 
 	plugs = {
 
@@ -84,7 +85,7 @@ Gaffer.Metadata.registerNode(
 		"user.*" : (
 
 			"deletable", True,
-			"labelPlugValueWidget:renameable", True,
+			"renameable", True,
 
 		),
 
@@ -129,13 +130,13 @@ class NodeUI( GafferUI.Widget ) :
 
 		return None
 
-	## Can be called to make the UI read only - must
-	# be implemented appropriately by derived classes.
+	## \deprecated
 	def setReadOnly( self, readOnly ) :
 
 		assert( isinstance( readOnly, bool ) )
 		self.__readOnly = readOnly
 
+	## \deprecated
 	def getReadOnly( self ) :
 
 		return self.__readOnly